float priorities
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@3057 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Base/GemMan.h
@@ -55,10 +55,10 @@ class GEM_EXTERN GemMan
   static void 	    initGem(void);
 
   //////////
-  static void 	    addObj(gemhead *obj, int priority);
+  static void 	    addObj(gemhead *obj, float priority);
     	
   //////////
-  static void 	    removeObj(gemhead *obj, int priority);
+  static void 	    removeObj(gemhead *obj, float priority);
 
   //////////
   // Is there a window.
